% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar.R
\name{bar}
\alias{bar}
\title{Draw bar plots.}
\usage{
bar(obj, ...)
}
\arguments{
\item{obj}{An object used to select a method.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object that can be combined with additional \code{geom_*}, \code{scale_*}, \code{theme_*}, etc.
}
\description{
The S3 method \code{\link[=bar]{bar()}} produces a bar plot to visually summarise the results of a simulation study analysed using \code{\link[=simsum]{simsum()}}.
}
\details{
Coverage, bias corrected coverage, and power will be based on the \eqn{\alpha} defined in the call to \code{simsum}.
}
