% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.multisimsum.R
\name{format.multisimsum}
\alias{format.multisimsum}
\title{Format multisimsum objects}
\usage{
\method{format}{multisimsum}(x, digits)
}
\arguments{
\item{x}{An object of class \code{multisimsum}.}

\item{digits}{Number of significant digits. Defaults to 4.}
}
\value{
An object of class \code{multisimsum} with its \code{summ} slot formatted for pretty printing.
}
\description{
Format an object of class \code{multisimsum} for pretty printing.
}
\seealso{
\code{\link[=multisimsum]{multisimsum()}}, \code{\link[=format.summary.multisimsum]{format.summary.multisimsum()}}
}
