% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.simsum.R
\name{format.simsum}
\alias{format.simsum}
\title{Format simsum objects}
\usage{
\method{format}{simsum}(x, digits)
}
\arguments{
\item{x}{An object of class \code{simsum}.}

\item{digits}{Number of significant digits. Defaults to 4.}
}
\value{
An object of class \code{simsum} with its \code{summ} slot formatted for pretty printing.
}
\description{
Format an object of class \code{simsum} for pretty printing.
}
\details{
format.simsum
}
\seealso{
\code{\link[=simsum]{simsum()}}, \code{\link[=format.summary.simsum]{format.summary.simsum()}}
}
