% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.multisimsum.R
\name{format.summary.multisimsum}
\alias{format.summary.multisimsum}
\title{Format summary.multisimsum objects}
\usage{
\method{format}{summary.multisimsum}(x, digits)
}
\arguments{
\item{x}{An object of class \code{summary.multisimsum}.}

\item{digits}{Number of significant digits. Defaults to 4.}
}
\value{
An object of class \code{summary.multisimsum} with its \code{summ} slot formatted for pretty printing.
}
\description{
Format an object of class \code{summary.multisimsum} for pretty printing.
}
\details{
format.summary.multisimsum
}
\seealso{
\code{\link[=multisimsum]{multisimsum()}}, \code{\link[=format.multisimsum]{format.multisimsum()}}
}
