% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.simsum.R
\name{format.summary.simsum}
\alias{format.summary.simsum}
\title{Format summary.simsum objects}
\usage{
\method{format}{summary.simsum}(x, digits)
}
\arguments{
\item{x}{An object of class \code{summary.simsum}.}

\item{digits}{Number of significant digits. Defaults to 4.}
}
\value{
An object of class \code{summary.simsum} with its \code{summ} slot formatted for pretty printing.
}
\description{
Format an object of class \code{summary.simsum} for pretty printing.
}
\details{
format.summary.simsum
}
\seealso{
\code{\link[=simsum]{simsum()}}, \code{\link[=format.simsum]{format.simsum()}}
}
