% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.miss.R
\name{get_data.miss}
\alias{get_data.miss}
\title{get_data.miss}
\usage{
\method{get_data}{miss}(x, ...)
}
\arguments{
\item{x}{An object of class \code{miss}.}

\item{...}{Ignored.}
}
\value{
A \code{data.frame} with estimates and standard errors substituted for a missingness boolean value.
}
\description{
Extract data from a miss object
}
\examples{
data("frailty", package = "rsimsum")
m <- miss(
  data = frailty, estvarname = "b", se = "se", par = "par",
  methodvar = "model", by = "fv_dist"
)
get_data(m)
}
\seealso{
\code{\link[=miss]{miss()}}, \code{\link[=get_data]{get_data()}}
}
