% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.summary.multisimsum}
\alias{is.summary.multisimsum}
\title{is.summary.multisimsum}
\usage{
is.summary.multisimsum(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Reports whether x is a summary.multisimsum object
}
