% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable.R
\name{kable.simsum}
\alias{kable.simsum}
\alias{kable.summary.simsum}
\alias{kable.multisimsum}
\alias{kable.summary.multisimsum}
\alias{kable}
\title{Create 'kable's}
\usage{
\method{kable}{simsum}(x, stats = NULL, digits = max(3, getOption("digits") - 3), ...)

\method{kable}{summary.simsum}(x, stats = NULL, digits = max(3, getOption("digits") - 3), ...)

\method{kable}{multisimsum}(x, stats = NULL, digits = max(3, getOption("digits") - 3), ...)

\method{kable}{summary.multisimsum}(x, stats = NULL, digits = max(3, getOption("digits") - 3), ...)

kable(x, ...)
}
\arguments{
\item{x}{An object of class \code{simsum}, \code{summary.simsum}, \code{multisimsum}, \code{summary.multisimsum};}

\item{stats}{Summary statistics to include. See \code{\link[=get_data]{get_data()}} for more details;}

\item{digits}{Maximum number of digits for numeric columns;}

\item{...}{Further arguments passed to \code{\link[knitr:kable]{knitr::kable()}}.}
}
\description{
Create tables in LaTeX, HTML, Markdown, or reStructuredText from objects of class \code{simsum}, \code{summary.simsum}, \code{multisimsum}, \code{summary.multisimsum}.
}
\seealso{
\code{\link[knitr:kable]{knitr::kable()}}
}
