% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{pattern}
\alias{pattern}
\title{Draw pattern plots.}
\usage{
pattern(obj, ...)
}
\arguments{
\item{obj}{An object used to select a method.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object that can be combined with additional \code{geom_*}, \code{scale_*}, \code{theme_*}, etc.
}
\description{
The S3 method \code{\link[=pattern]{pattern()}} produces a pattern plot to visualise estimates against standard errors for different methods.
}
\details{
\code{pattern()} requires \code{simsum} and \code{multisimsum} objects top be fit with the \code{x = TRUE} option.
}
