% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dropbig.multisimsum.R
\name{print.dropbig.multisimsum}
\alias{print.dropbig.multisimsum}
\title{print.dropbig.multisimsum}
\usage{
\method{print}{dropbig.multisimsum}(x, ...)
}
\arguments{
\item{x}{An object of class \code{dropbig.multisimsum}.}

\item{...}{Ignored.}
}
\description{
Print method for dropbig.simsum objects
}
\seealso{
\code{\link[=dropbig.multisimsum]{dropbig.multisimsum()}}
}
