% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dropbig.simsum.R
\name{print.dropbig.simsum}
\alias{print.dropbig.simsum}
\title{print.dropbig.simsum}
\usage{
\method{print}{dropbig.simsum}(x, ...)
}
\arguments{
\item{x}{An object of class \code{dropbig.simsum}.}

\item{...}{Ignored.}
}
\description{
Print method for dropbig.simsum objects
}
\seealso{
\code{\link[=dropbig.simsum]{dropbig.simsum()}}
}
