% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.miss.R
\name{print.miss}
\alias{print.miss}
\title{print.miss}
\usage{
\method{print}{miss}(x, fmt = "\%.4f", ...)
}
\arguments{
\item{x}{An object of class \code{miss}.}

\item{fmt}{Format string passed to \code{\link[base:sprintf]{base::sprintf()}} to format proportions of missing values when printed.}

\item{...}{Ignored.}
}
\description{
Print method for \code{miss} objects
}
\examples{
library(rsimsum)
data("frailty", package = "rsimsum")
m <- miss(
  data = frailty, estvarname = "b", se = "se", par = "par",
  methodvar = "model", by = "fv_dist"
)
print(m)
print(m, fmt = "\%.2f")
}
