% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.multisimsum.R
\name{print.multisimsum}
\alias{print.multisimsum}
\title{Print method for multisimsum objects}
\usage{
\method{print}{multisimsum}(x, ...)
}
\arguments{
\item{x}{An object of class \code{multisimsum}.}

\item{...}{Ignored.}
}
\description{
print.multisimsum
}
\examples{
data(frailty)
ms <- multisimsum(data = frailty, par = "par", true = c(trt = -0.50,
   fv = 0.75), estvarname = "b", se = "se", methodvar = "model",
   by = "fv_dist")
ms
}
