% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.simsum.R
\name{print.simsum}
\alias{print.simsum}
\title{Print method for simsum objects}
\usage{
\method{print}{simsum}(x, ...)
}
\arguments{
\item{x}{An object of class \code{simsum}.}

\item{...}{Ignored.}
}
\description{
print.simsum
}
\examples{
data("MIsim")
x = simsum(data = MIsim, estvarname = "b", true = 0.5, se = "se", methodvar = "method", mcse = TRUE)
x
}
