% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{relhaz}
\alias{relhaz}
\title{Example of a simulation study on survival modelling}
\format{A data frame with 8,000 rows and 6 variables:
\itemize{
\item \code{dataset} Simulated dataset number.
\item \code{n} Sample size of the simulate dataset.
\item \code{baseline} Baseline hazard function of the simulated dataset.
\item \code{model} Method used (\code{Cox} or \code{RP(2)}).
\item \code{theta} Point estimate for the log-hazard ratio.
\item \code{se} Standard error of the point estimate.
}}
\usage{
relhaz
}
\description{
A dataset from a simulation study assessing the impact of misspecifying the baseline hazard in survival models on regression coefficients. One thousand datasets were simulated, each containing a binary treatment variable with a log-hazard ratio of -0.50. Survival data was simulated for two different sample sizes, 50 and 250 individuals, and under two different baseline hazard functions, exponential and Weibull. Consequently, a Cox model and a Royston-Parmar model with two degrees of freedom were fit to each simulated dataset.
}
\details{
relhaz
}
\examples{
data("relhaz", package = "rsimsum")
}
\keyword{datasets}
