% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsimsum-package.R
\docType{package}
\name{rsimsum-package}
\alias{rsimsum}
\alias{rsimsum-package}
\title{rsimsum: Analysis of Simulation Studies Including Monte Carlo Error}
\description{
Summarise results from simulation studies and compute Monte Carlo standard errors of commonly used summary statistics. This package is modelled on the 'simsum' user-written command in 'Stata' (White I.R., 2010 \url{https://www.stata-journal.com/article.html?article=st0200}), further extending it with additional performance measures and functionality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ellessenne.github.io/rsimsum/}
  \item \url{https://github.com/ellessenne/rsimsum}
  \item Report bugs at \url{https://github.com/ellessenne/rsimsum/issues}
}

}
\author{
\strong{Maintainer}: Alessandro Gasparini \email{alessandro@ellessenne.xyz} (\href{https://orcid.org/0000-0002-8319-7624}{ORCID})

Authors:
\itemize{
  \item Ian R. White
}

}
\keyword{internal}
