% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsimsum.R
\name{rsimsum}
\alias{rsimsum}
\title{\code{rsimsum} package}
\description{
Summarise results from simulation studies and compute Monte Carlo standard errors of commonly used summary statistics. This package is modelled on the 'simsum' user-written command in 'Stata' (See White I.R., 2010 \url{http://www.stata-journal.com/article.html?article=st0200}). See the README on \href{https://github.com/ellessenne/rsimsum#readme}{GitHub}.
}
