% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsum.R
\name{simsum}
\alias{simsum}
\title{simsum}
\usage{
simsum(data, estvarname, true, se, methodvar = NULL, ref = NULL,
  df = NULL, dropbig = FALSE, max = 10, semax = 100, level = 0.95,
  by = NULL, sanitise = TRUE, mcse = FALSE, robust = FALSE,
  modelsemethod = "rmse")
}
\arguments{
\item{data}{A \code{data.frame} in which variable names are interpreted.}

\item{estvarname}{a variable containing the point estimates}

\item{true}{gives the true value of the parameter.  This is used in calculations of bias and coverage and is required whenever these performance measures are requested.}

\item{se}{lists the names of the variables containing the standard errors of the point estimates.  For data in long format, this is a single variable.}

\item{methodvar}{identifies the method}

\item{ref}{specifies the reference method against which relative precisions will be calculated.}

\item{dropbig}{specifies that point estimates or standard errors beyond the maximum acceptable values should be dropped.}

\item{max}{specifies the maximum acceptable absolute value of the point estimates, standardised to mean 0 and SD 1. The default value is 10.}

\item{semax}{specifies the maximum acceptable value of the standard error, as a multiple of the mean standard error.  The default value is 100.}

\item{level}{specifies the confidence level for coverages and powers. Default is 0.95.}

\item{by}{computes performance measures by varlist. Multiple factors are collapsed into a single one using \code{\link[base:interaction]{base::interaction()}}}

\item{sanitise}{Sanitise column names passed to \code{simsum} by removing all dot characters (\code{.}).}

\item{mcse}{reports Monte Carlo standard errors for all performance measures.}

\item{robust}{is only useful if mcse is also specified.  It requests robust Monte Carlo standard errors for the performance measures \code{empse}, \code{relprec} and \code{relerror}, instead of those based on an assumption of normally distributed point estimates.}

\item{modelsemethod}{specifies whether the model standard error should be computed as the root mean squared value (the default) or as the arithmetic mean.}

\item{perfmeasures}{If none of the following options is specified, then all available performance measures are computed. Possible values are:
\itemize{
\item \code{bsims} reports the number of simulations with non-missing point estimates.
\item \code{sesims} reports the number of simulations with non-missing standard errors.
\item \code{bias} estimates the bias in the point estimates.
\item \code{empse} estimates the empirical standard error -- the standard deviation of the point estimates.
\item \code{relprec} estimates the relative precision -- the inverse squared ratio of the empirical standard error of this method to the empirical standard error of the reference method.  This calculation is slow:     omitting it can reduce run time by up to 90%.
\item \code{mse} estimates the mean squared error.
\item \code{modelse} estimates the model-based standard error. See modelsemethod() above.
\item \code{relerror} estimates the proportional error in the model-based standard error, using the empirical standard error as gold standard.
\item \code{cover} estimates the coverage of nominal confidence intervals at the specified level.
\item \code{power} estimates the power to reject the null hypothesis that the true parameter is zero, at the specified level.
}}
}
\value{
An object of class \code{simsum}.
}
\details{
Beware of \code{by} or \code{methodvar} whose levels contain dots (\code{.})!
}
\examples{
\dontrun{
x = simsum(...)
}
}
