% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.simsum.R
\name{summary.simsum}
\alias{summary.simsum}
\title{Summarising simsum objects}
\usage{
\method{summary}{simsum}(x, level = 0.95)
}
\arguments{
\item{x}{An object of class \code{simsum}.}

\item{level}{Significance level for confidence intervals based on Monte Carlo standard errors. Ignored if a \code{simsum} object is obtained with \code{mcse = FALSE}.}
}
\value{
An object of class \code{summary.simsum}.
}
\description{
The \code{summary()} method for objects of class \code{simsum} returns confidence intervals for performance measures based on Monte Carlo standard errors.
}
\details{
summary.simsum
}
\examples{
data("MIsim")
x = simsum(data = MIsim, estvarname = "b", true = 0.5, se = "se", methodvar = "method", mcse = TRUE)
xs = summary(x)
xs
}
\seealso{
\code{\link[=simsum]{simsum()}}, \code{\link[=print.summary.simsum]{print.summary.simsum()}}
}
