% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip}
\alias{zip}
\title{Draw zip plots.}
\usage{
zip(obj, ...)
}
\arguments{
\item{obj}{An object used to select a method.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object that can be combined with additional \code{geom_*}, \code{scale_*}, \code{theme_*}, etc.
}
\description{
The S3 method \code{\link[=zip]{zip()}} produces a zip plot to visualise coverage. Zip plots are introduced in Morris \emph{et al}., 2017.
}
\references{
Morris, T.P, White, I.R. and Crowther, M.J. 2017. Using simulation studies to evaluate statistical methods. \href{https://arxiv.org/abs/1712.03198}{arXiv:1712.03198}
}
