% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropbig.R
\name{dropbig}
\alias{dropbig}
\title{Return observations dropped by a function}
\usage{
dropbig(x, ...)
}
\arguments{
\item{x}{An object used to select a method.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
\code{dropbig} returns observations dropped by a function while performing some calculation - for any reason.
}
\details{
dropbig
}
