% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Extract data slot from objects}
\usage{
get_data(x, ...)
}
\arguments{
\item{x}{An object used to select a method.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
\code{get_data} returns a data set contained in an S3 object.
}
