% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.simsum.R
\name{get_data.simsum}
\alias{get_data.simsum}
\title{Extract data from a simsum object}
\usage{
\method{get_data}{simsum}(x, ...)
}
\arguments{
\item{x}{An object of class \code{simsum}.}

\item{...}{Ignored.}
}
\value{
A \code{data.frame} containing summary statistics from a simulation study.
}
\description{
get_data.simsum
}
\examples{
data(MIsim)
x <- simsum(data = MIsim, estvarname = "b", true = 0.5, se = "se",
            methodvar = "method", mcse = TRUE)
get_data(x)
}
\seealso{
\code{\link[=simsum]{simsum()}}, \code{\link[=get_data]{get_data()}}
}
