% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.summary.simsum.R
\name{get_data.summary.simsum}
\alias{get_data.summary.simsum}
\title{Extract data from a summary.simsum object}
\usage{
\method{get_data}{summary.simsum}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.simsum}.}

\item{...}{Ignored.}
}
\value{
A \code{data.frame} containing summary statistics from a simulation study.
}
\description{
get_data.summary.simsum
}
\examples{
data(MIsim)
x <- simsum(data = MIsim, estvarname = "b", true = 0.5, se = "se", methodvar = "method", mcse = TRUE)
xs <- summary(x)
get_data(xs)
}
\seealso{
\code{\link[=summary.simsum]{summary.simsum()}}, \code{\link[=get_data]{get_data()}}
}
