% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsimsum-package.R
\docType{package}
\name{rsimsum}
\alias{rsimsum}
\alias{rsimsum-package}
\title{Analysis of Simulation Studies Including Monte Carlo Error}
\description{
Summarise results from simulation studies and compute Monte Carlo
standard errors of commonly used summary statistics. This package is modelled
on the 'simsum' user-written command in 'Stata' (See White I.R., 2010
\url{http://www.stata-journal.com/article.html?article=st0200}).
}
\author{
Alessandro Gasparini (alessandro.gasparini@ki.se)
}
